deployment_package/requirements.txt
# Core Flask and Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Caching==2.1.0
Werkzeug==2.3.7

# Data Processing and Excel
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1

# Document Processing
python-docx==0.8.11
docxtpl==0.16.7
docxcompose==1.4.0
lxml==4.9.3

# Image Processing
Pillow==10.1.0
qrcode==7.4.2

# Database
psycopg2-binary==2.9.9

# Utilities
python-dateutil==2.8.2
pytz==2023.3
jellyfish==1.2.0
requests>=2.32.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.27.0

# Development and Testing (optional)
# pytest==7.4.3
# pytest-cov==4.1.0

# Python version compatibility
# Requires Python 3.8+


